fix copy on mac
